requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
tiktoken
sentence-transformers
huggingface-hub